requirements.txt
Flask~=3.1.0
firebase-admin~=6.7.0
Flask-OAuthlib
Flask-SQLAlchemy
pymysql
streamlit
groq~=0.22.0
psycopg2
SQLAlchemy~=2.0.40
googlesearch-python